// adopt-me-front/src/types/ActivityType.ts
export enum ActivityType {
    LOGIN_SUCCESS = "LOGIN_SUCCESS",
    LOGIN_FAILED = "LOGIN_FAILED",
    ADDED_ATTACHMENT = "ADDED_ATTACHMENT",
    REMOVED_ATTACHMENT = "REMOVED_ATTACHMENT",
    ADDED_ANIMAL = "ADDED_ANIMAL",
    UPDATE_ANIMAL = "UPDATE_ANIMAL",
    REMOVED_ANIMAL = "REMOVED_ANIMAL",
    ADDED_TAG = "ADDED_TAG",
    REMOVED_TAG = "REMOVED_TAG",
    ADDED_USER = "ADDED_USER",
    REMOVED_USER = "REMOVED_USER",
    SENT_NOTIFICATION = "SENT_NOTIFICATION",
}